requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
cv2-python==4.8.1.78
opencv-python==4.8.1.78
numpy==1.24.3
pydantic==2.5.0
pydantic-core==2.14.1
pydantic-settings==2.1.0
starlette==0.27.0
pyyaml==6.0.1
click==8.1.7
h11==0.14.0
typing-extensions==4.8.0
anyio==4.1.0
sniffio==1.3.0
httptools==0.6.1
python-dotenv==1.0.0
insightface==0.7.3
onnxruntime==1.17.0
onnx==1.15.0
scikit-image==0.22.0
scipy==1.11.4
paho-mqtt==1.6.1
